requirements.txt
aiogram==2.25.1
pyttsx3==2.90
moviepy==1.0.3
numpy==1.24.2
openai==0.27.2
pydub==0.25.1
imageio==2.26.1
SQLAlchemy==2.0.7
Pillow==9.4.0
pytest==7.2.2
torch==2.0.0
transformers==4.27.4
asyncpg==0.27.0
google==3.0.0
googletrans==4.0.0rc1
